+++ resources/js/Components/Pages/Welcome/Register.vue
@@ -81,6 +81,14 @@
               />
             </div>
 
+            <div class="invite-code">
+              <div class="mt-4 ml-4 border-t border-t-gray-200 text-center font-semibold">
+                <div @click.prevent="appSettingStore.btnRedirect(`subscribe`)"
+                     class="text-blue-500 hover:text-blue-400 hover:cursor-pointer hover:scale-105 transition">
+                  👉 Need an invite code? <span class="underline">Click here</span></div>
+              </div>
+            </div>
+
             <!--            Jetstream/Fortify Multi-Auth: Roles, Permissions and Guards-->
             <!--            https://www.youtube.com/watch?v=NiQSNjWKLfU-->